requirement.txt
tensorflow
numpy
opencv-python
streamlit
matplotlib
seaborn
scikit-learn
pillow